Part 3 Maybe Final Commit
This guy works. When you run it, it loaded map of the U.S. and painted
every Starbucks location based on population change per Starbucks. When
you click the window, it highlights the neareast Starbucks. Actually
there are still small problems, but it does what the requirements want
us to do.

It is not as beautiful as others (actually it is ugly), but it paints
based on population change per Starbucks. I already tried my best...do
not have strength to do anything else.