Part 3 Final Commit
Works...ugly, not pretty at all, but it works. When you run it, it loads
map of the U.S. and paints all Starbucks locations based on population
change per Starbucks. When you click, it highlights the nearest
Starbucks.

It is very very ugly and not interesting; the map and Starbucks do not
match each other, and the color changes in a wired way. But I already
tried my best...have no strength to do anything else. This is what I am
done.